Decouple rating command handling from command source
